Further technical graphics

+++ code/game/objects/items/wall_mounted.dm
@@ -9,6 +9,7 @@
 	var/result_path
 	var/wall_external = FALSE // For frames that are external to the wall they are placed on, like light fixtures and cameras.
 	var/pixel_shift //The amount of pixels
+	var/multi_use = FALSE //BUBBER EDIT ADDITION - lewd portals are just wallframes that are placed in sets.
 
 /obj/item/wallframe/proc/try_build(turf/on_wall, mob/user)
 	if(get_dist(on_wall,user) > 1)
@@ -53,8 +54,10 @@
 				if(WEST)
 					hanging_object.pixel_x = -pixel_shift
 		after_attach(hanging_object)
-
-	qdel(src)
+	//BUBBER EDIT START - For lewd_portals, you can place multiple with the same frame.
+	if(!multi_use)
+		qdel(src)
+	//BUBBER EDIT END
 
 /obj/item/wallframe/proc/after_attach(obj/attached_to)
 	transfer_fingerprints_to(attached_to)
